+++ CHANGES.md
@@ -2,6 +2,37 @@
 
 ## cordova-sqlite-evplus-feat-android-db-location 0.0.3-dev
 
+## cordova-plugin-sqlite-evplus-ext-common-free 0.8.0-rc02
+
+- SQLite 3.36.0 update from cordova-sqlite-evplus-ext-free-dependencies@0.4.1
+
+## cordova-plugin-sqlite-evplus-ext-common-free 0.8.0-rc01
+
+- use cordova-sqlite-evplus-ext-free-dependencies 0.4.0 with BLOBFROMBASE64 feature
+
+## cordova-plugin-sqlite-evplus-ext-common-free 0.7.0
+
+- fix condition used to close database on Android
+
+### cordova-sqlite-evcore-common-free 0.6.0-dev
+
+- update Windows minimum & target platform versions - BREAKING CHANGE
+
+## cordova-plugin-sqlite-evplus-ext-common-free 0.6.0-pre01
+
+- updates from cordova-sqlite-evplus-ext-free-dependencies@0.3.0
+  - update evplus NDK library - solution for Android 11 with target SDK 30
+  - SQLite 3.35.5 update for all platforms
+
+## cordova-plugin-sqlite-evplus-ext-common-free 0.5.0-pre01
+
+- updates from cordova-sqlite-evplus-ext-free-dependencies 0.2.0-rc01
+  - BREAKING: drop Android pre-5.1 support
+  - update SQLite to 3.35.4
+  - fix for emojis and other 4-byte UTF-8 characters on Android
+  - update affected test cases
+  - update documentation
+
 ## cordova-sqlite-evplus-feat-android-db-location 0.0.2-dev
 
 ## cordova-sqlite-evcore-feat-android-db-location 0.8.0